Updated openstack/openstack Project: openstack/neutron  76df414d7a0a63577b0cc08e84ba096bea2b391f
Watch for 'new' events in ovsdb monitor for ofport

This adjusts the event handling logic in the ovs interface
monitor to watch for 'new' events as well because this is
when ofports are assigned to the interface. Previously we
were only watching for 'insert' events, which would never
include the ofport so this led to every port being skipped
initially by the OVS agent and then picked up on the next
polling interval.

Closes-Bug: #1560464
Change-Id: I46ac508a839c6db779323d5afb083f3425f96e87