Updated openstack/openstack Project: openstack-infra/shade  c7c48c9d05a316646d069c00060613cfc41041fd
Create clouds in Functional Test base class

We need to use the long-form of the OpenStackCloud constructor to enable
the internal_exception logging. In order for that not to be crazy, we
should do it in a base TestCase class. Also, that way we've got
consistently named clouds in all of the tests.

Change-Id: I42da312dbcaa8926d28e97e37d13370a544cf4bd