Updated openstack/openstack Project: openstack-infra/system-config  03d09b8af4d8f5d6451baff6db94f09188594c23
Don't set the host's name as UUID

It looks like we solved the duplicate server problem twice in
conflicting ways. Using uuid in the inventory is not needed, bcause
we're making a specific inventory for the ansible commands and avoiding
the OpenStack inventory. So the ansible run has no idea of any other
servers other than the one we're making right now. With that, we can use
name as the hostname rather than UUID.

Story: 2000520
Change-Id: Idb967e10fc00471923077e4e9caa32fdb4c1cc78