Updated openstack/openstack Project: openstack/nova  1906aa421433713c2719dcd1ce57c541fb46fbd2
Add a DatabasePoisonFixture

This is the beginning of a poison fixture for tests that don't use
the database. If they do, we interrupt them deep in oslo_db, log a
warning, and return a MagicMock. This keeps them from breaking
global database state, but doesn't actually poison them. This should
be followed-up with a patch to raise an exception, and fix the rest
of the tests that depend on the broken behavior.

Since some tests claim to not be DB test, but use the fixture
themselves anyway, this adds a USES_DB_SELF flag to help gate
the application of the poison.

Change-Id: If071a12a33aad86e918a9f441a246a930e921c8d