Updated openstack/openstack Project: openstack/ironic  e3c2d10eed1c69e260ef65b92f40acacff26a944
Fix sphinx docs build

When building docs with sphinx, the objects.register_all() code path is
not necessarily called, and therefore objects.* is not necessarily
imported before it is used. This throws an error while building the
sample ChassisCollection response.

If objects.Chassis is not available, call register_all() before using
it.

Change-Id: Ic0041085639f3d5a900ad54c2637c77aa467ba35
Closes-Bug: #1560508