Updated openstack/openstack Project: openstack/django_openstack_auth  27d9604ac091b1a97216a2bb21c939151bcd0087
Don't call the Keystone client if the token is None

There are certain cases the user.Token constructor is called with
the default unscoped_token value of None. When this happens using
PKI, an unexpected error occurs. If the token is None, we know it's
not pki and should not make an extra call to the Keystone client to
check.

A test is provided to demonstrate the issue, and the test fails
without the included code change.

Change-Id: Ie8e93eda4cf8e8285785335126624781470be19c
Closes-Bug: #1551977