Updated openstack/openstack Project: openstack/governance  7210e57fa011c2f987532da2f47527b791aac213
Add testing requirement to assert:supports-upgrade tag

The supports-upgrade tag is missing a requirement for actual testing,
and as a result, a couple of projects have the tag applied which do
not have any upgrade testing. The rolling-upgrade tag has a requirement
for testing, and this one should too.

This also removes the supports-upgrade tag from aodh and gnocchi, which
do not appear to have any grenade (or similar) jobs configured.

Change-Id: I47457ecd5ceeac2dd3e847c887fb43fafe97c60b