Updated openstack/openstack Project: openstack-infra/storyboard-webclient  69d5e837e60bfe58b1db16dbb38503d0c7461f14
Stop updating the board in-place when moving lanes

It was possible to move a lane, then pick up another before the
update was complete. The second lane would then remain on the screen
following the cursor until the page was refreshed. This commit
fixes that bug by not changing the client-side model when the
PUT request completes.

Change-Id: Id4e5067c6c9fc080eae12149e9eeff860846243b