Updated openstack/openstack Project: openstack/ironic  a188a313650266cb3103e60702d0505557d8ea88
Devstack: add check of chassis creating

Current devstack scripts will not check if chassis creating
is successfully or not, just go ahead to run next actions,
so when we use chassis_uuid, if it is empty, will break the
devstack, but user don't have idea why missing this chassis_uuid.

With this patch, we add code to check if chassis creating
is successfully or not, and exit the devstack if failed.

Change-Id: I575578a05274fff0cffbb86e8780a8ef5a6366b6
Closes-Bug: #1558340