Updated openstack/openstack Project: openstack-infra/system-config  453051aafbc423143be4de922688ab90284528d7
Don't pass OS_CLOUD and OS_REGION_NAME to expand-groups

There is a bug in OCC that causes an envvars cloud to be created when
the only two env vars are the selectors OS_CLOUD and OS_REGION_NAME. So
exclude them from the envionment when running the group creation
command.

Also, there is a bug in the invocation of the hostname playbook, in that
it was passing in the UUID as the target to run against, but we're
writing out a name-based inventory.

Change-Id: I0b524dc43ec96c6645ae82a090744eab463e7fb9