Updated openstack/openstack Project: openstack-infra/project-config  e0c6d29fff27ae822075c73b69dd3597ec626890
Refactor the v3 only job

Refactor the identity v3 only job that was being run against devstack so
that it can be applied to additional services. The keystone team has
been trying to move people off the V2 API for a while now and this job
has been successfully running on the devstack gate for a while now, but
as is does not give us any indication of the readiness of other
services. We need to run this against individual services.

As part of this we change the job from always non-voting to using the
-nv suffix so that we can start making the job voting on a per-service
basis as services get fixed.

Change-Id: I90e7fdd93f243f2568370f084001531c50cb080e