Updated openstack/openstack Project: openstack-infra/project-config  eb5242d5bec7473f7a7bde1a747f8c1974bbdc39
trove: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I48b1ef7211151ba1e233c6196b5caab96fbf03d4