Updated openstack/openstack Project: openstack-infra/project-config  5a05fa2ccc4f115b286b564a48cbe9955fe5d6f0
sqlalchemy-migrate: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I06eed49d63cef12169195d234a5ff0e4682def06