Updated openstack/openstack Project: openstack/oslo.messaging  3802dd5e4b0dba554651870035365e466f36220c
Always set all socket timeouts

When exchange/queue are declared the timeout is the default (None)
when it should be the no more than heartbeat timeout.

Also we have to set TCP_USER_TIMEOUT to ensure our timeouts are
respected. socket.settimeout() of python relies on select() that can
be blocked because kernel doesn't return during a recv() because of
TCP connection breakage.

This change fixes that by always setting TCP_USER_TIMEOUT and python
socket timeout each time the connection is establish.

Change-Id: Ibf6139ac2c22d9eeda7030fb87b7f1139d92332e