Updated openstack/openstack Project: openstack-infra/project-config  0898df1ccfbf41a8a27366b44d6d1abf9a5088fd
Creation of a dedicated repo for API and media storage client

Currently the related freezer code to talk to the API and
query the storage media is hosted in the openstack/freezer repo.
We would like to follow the same convention/organization as
other OpenStack projects and place the code on a dedicated repo,
hoping also to reduce complexity and increase readability.

The new repo name would be: openstack/python-freezerclient

Implements blueprint: freezerclient

Needed-By: I0d2ca6623c81185ab8443ade7d0234e4ded63d7f
Change-Id: I8441d5bf0e3dcbb679eb39ad29bf89cd957f37c4