Updated openstack/openstack Project: openstack-infra/release-tools  9ee1cb3c08693e467eccf51e0cfc2aac779f8b35
Added tool to filter out bugs with specific tag

Sometimes you have a list of bugs, and you need to filter out bugs that
have some specific tag. For example, if you consider bugs for backports
to a branch, you probably want to skip those bugs that already have
in-stable-<release> tag set.

And that's the use case for the filtering tool.

The tool can be combined with other tools that output the list of bug
numbers, like bugs-fixed-since.py or lp-filter-bugs-by-importance.py
using standard Unix pipelines.

Change-Id: I66e84b175c2a4efc87d897ac6d861c38b2ddfb3d