Updated openstack/openstack Project: openstack-infra/puppet-openstackid  d738f5647d83d1dfbc2362224cb855654f9e7616
Default to hostname not fqdn for laravel env

Laravel expects the short hostname for its environment identifier,
but we didn't notice this before because production openstackid.org
and openstackid-dev.openstack.org had incorrect /etc/hostname files
containing their fully qualified hostnames rather than their short
hostnames. Now that we correctly set the contents of the hostname
file in new deployments, this will break without correcting the
Puppet module.

The plan is to merge this with Puppet disabled for openstackid.org,
confirm that correcting the hostname on openstackid-dev revives it,
then manually apply the same correction and this patch on production
openstackid.org before reenabling Puppet there.

Change-Id: I41ddb3cd12fa564d19cedc18ca5585d2ca7481da
Depends-On: If67783a2cbaf77e89a3d2aae80a9206b547dbac6