Updated openstack/openstack Project: openstack-infra/storyboard-webclient  27e0a4b19dd9371990b360b1f73b5013f6410baf
Allow an update of the SearchCriteriaController to be triggered

This commit makes the `refresh-types` event call the init() function of
SearchCriteriaController, and also changes the init() function to check
for an array of types set in $scope.

This allows the set of types to be changed from outside the controller.

Change-Id: Id2bc257a72a62d787499cf079a7a3c44a6ceca36