Updated openstack/openstack Project: openstack-infra/project-config  28c05ecc7661df626268627e5670b0621e5625ef
subunit2sql: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I1e811361a2508c4e2ce940668792304814137644