Updated openstack/openstack Project: openstack/neutron-lbaas  3a3bee9c78b8601750c3fe9b005a58273f6daee3
py3: Replace dict.iteritems() with dict.items()

While six.iteritems(dict) was preferred before, it was decided to use
dict.items() to make the code more readable. It was considered that
the overhead of the creation of a temporary list on Python 2 is
negligible. If it's not negligible, it's an hint for a design issue.

http://lists.openstack.org/pipermail/openstack-dev/2015-June/066391.html

Blueprint: neutron-python3
Change-Id: I8a77b06ed36e2dcadfc3c408c7998982f4e521b7