Updated openstack/openstack Project: openstack-infra/project-config  74555e0dea1cfcac99abf7cbc1bd274a7d7d9d68
Stop using rax for multinode

Make this change until we have working eth0 and eth1 interfaces on rax
VMs as the multinode tests expect the private network to function if it
is present.

The underlying issue is that eth1 attempts to configure itself iwth dhcp
but there is no dhcp in rax and the glean config isn't in place before
eth1 is up'd. A new release of dib should fix this by removing the eth1
dhcp config and forcing glean to configure this interface.

Change-Id: I4daa46c56d701aedd99e10a69e7fa09b40dfaa47