Updated openstack/openstack Project: openstack-infra/project-config  e92a79e11c115a6148ea304a02adde9a5669d935
Allow test slaves to talk to themselves

Update iptables rules to allow slaves to talk to themselves over their
known IP addresses. This should allow services to listen on public or
private addresses in addition to the loopback address and talk to those
services over these IPs.

Specifically this should address a problem with rabbitmq coming up on
the public IP but not being able to talk to it due to iptables.

Change-Id: If851eb8ce97ca0c0c113f0a924ec0916c993cf87