Updated openstack/openstack Project: openstack/zaqar  0e1f4aed2846ead5fe47484a88d88a0979bf1bd6
Warn on upcoming deprecations for v1.0 and v1.1

v1.0 has been deprecated for a while now but it's still on by default.
V1.1, on the other hand, is still supported but on its way to being
deprecated. Warn users that both APIs will be turned off by default
during the Newton cycle.

Change-Id: Ibd6e12bd9a8ba414bb9ea26136a5acc1ca092142