Updated openstack/openstack Project: openstack-infra/system-config  9dc0c981eaf45bdd54b8f7dc74c75f5df4a466fd
Adding meetbot and statusbot support for openstack-product

The openstack-product IRC channel will be used by the OpenStack Product
Working Group for additional conversations when the meetings take longer
than our allocated time on the openstack meeting channel.  The product
working group will also be in the channel so that we can answer any
questions related to our working group and/or user stories.

Change-Id: I26b1304a16bb4ce6892b2407b3297c8b48d3bf58