Updated openstack/openstack Project: openstack/nova  fdc2c943e8846a09fb6f22acb4003358ecafc9b9
Make FlavorList.get_all() return results from the API and main DBs

This makes the FlavorList.get_all() method return a merged result
of things in the main database and the API one. We favor the API DB
when we have been given a limit, and avoid a call to the main DB if
it makes sense. We also trim the limit when we make the second call,
as needed to avoid returning too many items.

Related to blueprint flavor-cell-api

Co-Authored-By: Lalit Dagre <lalitdagre@gmail.com>
Change-Id: If9de62cf0470e046dec10f0b65d00dd29d4d7ddd