Updated openstack/openstack Project: openstack-dev/pbr  b4d21581e267e821d5ac524b4dfc63c149ffcfd5
Do not convert git tags when searching history

Currently, when the sem-ver processing is searching the git history,
it stores the canonical version and uses that as the tag, which is
fine if the project uses the canonical version as a tag. Instead,
return the original tag so that history can be correctly checked.

Change-Id: I1b83fc3dc3dd287ae3b23fd4cba07fdfdd673a7f
Closes-Bug: 1561254