Updated openstack/openstack Project: openstack/nova  caa830503058876038809aeb04bae06e66463aed
Try to repopulate instance_group if it is None

There are cases that we need to create new RequestSpect object and we
pass None for instance_group argument. It might cause server group
policies to be omitted during filtering phase. Therefore if
instance_group is None we should try to repopulate instance_group
basing on data that is in filter_properties.

Change-Id: Id7e669e0a6db1ff1052c42006f1a141bdb8cdf29
Closes-Bug: #1561022