Updated openstack/openstack Project: openstack/diskimage-builder  81633f69c85d3ff3950016dba1f5ed172f71586e
add pkg-map to pip-and-virtualenv element

Add a pkg-map file so that Gentoo can avoid polluting the FS with
files not tracked in the package manager.

Changed the directory name so that the source-install method would
actually be disabled when DIB_INSTALLTYPE_pip_and_virtualenv=package
is set.  Also changed source-repository line to be consistent with
the element name.

Change-Id: I5225bd7113ff2588c755ad4cd4ae9a2d78fff196