Updated openstack/openstack Project: openstack/tripleo-heat-templates  1697dc1a54ed8daa43e346bd563de9ae9a051710
Remove unused user resource.

The user is created by installation of of the pacemaker package, so it's
not required to add it to the resource catalog.

This paves the way to merge the refeactoring of the puppet-pacemaker
module[1].  It brings a lot of changes, one of them is an idempotent
handling of the hacluster user's password.  Removing it here prevents
duplicate resource error durring puppet run.

[1] https://review.openstack.org/#/c/294182/

Change-Id: I56849d9fc00bd3ce342d5c440cfe7c5b6d26b5bf