Updated openstack/openstack Project: openstack-infra/project-config  a62d5e3cb1725b16b3cfe559e1c48fd7baf20c0e
tuning-box: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I71e2f478577cd3ac2776ab242bdfd91173f9f437