Updated openstack/openstack Project: openstack-infra/project-config  e01d3f0515c140c8e9df73f31a899f28d17aa79f
fuel-ostf: use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
These repo need databases setup for tests.

Change-Id: I9448b0be4562d0aee92527b09143e04455c4b8c4