Updated openstack/openstack Project: openstack/neutron  8c4e7ab63ae518b27b0e822bedbef159132d53eb
Fix fullstack LB agent not connecting to rabbit

Using the broadcast address of a network is not going to work
very well.

It's still possible for different LB agents (From the same test)
to pick the same address, and that will be handled in a follow up
via some sort of atomic resource allocation mechanism.

Change-Id: I666cf7dbe5dd6a0eaf4130ee1e52cf29edb2ddca
Related-Bug: #1561248