Updated openstack/openstack Project: openstack/nova  b27b063adb7fe19fda8c0c3dd1b9be39465111f8
Make Flavor access routines work against API database

This makes the access-related bits of the Flavor object work when the
flavor is in the API database instead of the main one. Included are
fixes for lazy-loading projects, add/remove access, and the projects
part of save(). In order to try to improve performance, this creates a
property on the object called "in_api" which will detect whether the
flavor is in the API database, caching the true case only so we can
stop checking. This is not a remotable field, which means it won't go
over RPC (and thus doesn't need a version bump), but since we only ever
do the access manipulation from the API, this is fine and we still get
the benefit.

Related to blueprint flavor-cell-api

Change-Id: I03a856235d653cf8831740a344072a897719bfa6