Updated openstack/openstack Project: openstack/ironic  c51a1510b8f871fc4f66fd8750b09bc1c44eab74
Config to stop powering off nodes on failure

This adds the config option deploy.power_off_after_deploy_failure,
which will prevent powering off nodes once a deployment fails.

Powering off nodes means loss of the log output from the agent.
Choosing to leave the machine powered on would give an operator
the opportunity to troubleshoot the issue using agent logs on the
system in case of failure, or to prevent needing to power the machine
back on for provisioning after cleaning.

As a note, Nova will still attempt to delete the node on a failed
deployment, so the option will not be very effective in non-standalone
deployments until a similar patch is put into nova.

Related-bug: 1531217
Change-Id: I7c25946de50ffda2869307cb4016e0ccfe81d192
Co-Authored-By: Josh Gachnang <josh@servercobra.com>