Updated openstack/openstack Project: openstack/trove  c0dc39e2af57b20493aef19dfee1d07345671706
Unhandled messages logged in tox tests

This is because the logger wasn't suitable masked (or the task manager
needs to properly configure logging). In the long term, that may be a
better solution, have to fix the fake taskmanager.py, maybe?

In the short term, mocking the logger is the best we have.

Change-Id: Idb27d581fda68d5129e12383f964a3ce6641a080
Closes-Bug: 1560760