Updated openstack/openstack Project: openstack/nova  e80376e3796cc20a400db423b828a98de6f528e6
Include CellMapping in InstanceMapping object

Right now InstanceMapping exposes the cell_mappings.id db field which is
just an implementation detail of the db relationship and should not be
exposed in the object. It is also useless as a way to look up the actual
CellMapping through that object interface. And InstanceMapping is only
looked up in order to find a CellMapping.

So the CellMapping should be joined during the db query and loaded on
the InstanceMapping object by default.

Change-Id: Ia8691b76bba310327bfe0995964525409794d1af