Updated openstack/openstack Project: openstack/ironic  8e15d6a1388f49026c4d35ba87219a49ccc3c134
Allow fetching IPA ramdisk with branch name

This adds branch name to the default URLs for IPA images, ensuring that
devstack runs will use the correct branch of IPA for tests.

It introduces a variable IPA_DOWNLOAD_BRANCH to accomplish this, which
defaults to master. This default will be set to the branch name on
stable branches, e.g. stable/mitaka. It is modified to match our URL
scheme for agent images, e.g. stable-mitaka.

Change-Id: I3ec6dafb9522b5fe79e0c6681c2578d2a723380a
Closes-Bug: #1561756