Updated openstack/openstack Project: openstack/trove  d6d212fe454ff82ec7a8e76715a7321d0d567d39
Fix replication scenario tests

When the change for slave_of -> replica_of was reverted, the
test calls using the python interface (int-tests) were also
reverted.  Unfortunately the scenario tests were not, so they
are now broken.

This fix synchronizes the calls to all use replica_of, which is
the preferred argument.

Closes-Bug: #1558584
Change-Id: I45c8c5a21c141ee7ebd8b5616bac174fe99eb907