Updated openstack/openstack Project: openstack/nova  9ad64ad7c2c6180d63acd80d5fd25f7a8e987610
Make flavor extra_specs operations work against the API DB

We already had loading extra_specs working from the API DB, thanks
to the SQLA relationship. This makes saving them work against it
as well.

This adds a test_save_api() copy of the test_save method with the
appropriate changes, which also increases coverage over the access
changes from a previous patch now that we can run the whole save()
method against API. Later we'll just delete the old test method
when we drop compatibility for saving to the old database.

Further coverage for this will be increased in a later patch,
when we gain the ability to create flavors in the API DB natively,
which will let us try to run more functional tests against flavors
in both databases the way we'd expect in production.

Related to blueprint flavor-cell-api

Change-Id: I63706bde5eaaa9af3b864307b446ed0cf30d8a1b