Updated openstack/openstack Project: openstack/nova  c21d2685478e19bb3a5cd230d6bc192a45320bc0
config options: Use OptGroup for listing options

Because of bug 1395819 in oslo.config we couldn't pass in the OptGroup
object when listing the options. We had to use the name of the object.
This got solved in version 3.2.0 of oslo.config.
As we use oslo.config >= 3.7.0 now, we can make use of the OptGroup
object itself and don't need the "workaround" anymore. This allows the
generation of the group help too.

Implements: bp centralize-config-options-newton

Change-Id: I1fe86c908127fbea72e6239277f4c0f49af1ff49