Updated openstack/openstack Project: openstack/neutron  c2d9c5c454c6a1b395d1bad20d16bd013e015605
firewall: don't warn about a driver that does not accept bridge

Drivers that do not need the bridge parameter are safe to ignore it.
Actually, the base abstract interface for those drivers does not allow
to pass the new parameter.

Ideally, the base class __init__ signature would change to accept the
argument, but we can't do it without breaking API.

So instead, just handle both types of drivers - those that accept the
additional argument, and those that don't. And don't assume the latter
are somehow wrong.

Change-Id: Iceee46f63669b28e3a34d207216d864f1bfa5cf8
Closes-Bug: #1561243