Updated openstack/openstack Project: openstack-infra/project-config  1c784099d2476561d2f02681d36bcf8b96a19e70
Stop emitting warning on missing nose files

OpenStack, with one exception, uses testr to run its tests. That means
that the expected outcome of run_tox.sh is that there is no
nose_results.html - however we emit a warning about it being missing on
every run. For folks less familiar with our systems trying to debug test
failures, it's a line that could confuse them.

Change-Id: I60f6b1f1e49bb8b001d5e9743273830188b150f1