Updated openstack/openstack Project: openstack/keystone  3406748a9b24187207f47f8d9a2b9b4064380d69
Update dev docs and sample script for v3/bootstrap

The default value for the admin_token parameter was removed from the
sample config file in ea952bd2, so the recommendation in the developer
documentation to use it without first setting it will result in an
error. This patch updates the developer documentation and the
sample_data.sh script to use the Identity v3 API and to follow the
latest recommendations for bootstrapping an admin user with which to
insert sample data.

Change-Id: I424ff6129d4ddcd63fb2bed4eabcbe910ab0153e