Updated openstack/openstack Project: openstack-infra/shade  23f5e995dad0d11ff5708336e36972479f9b236d
Add wait support to create_image_snapshot()

When creating a snapshot from a server sometimes it is desired to make
this a synchronous operation. This is the same reasons we have a wait
on the normal image create. This commit adds the necessary args and
logic to the create_image_snapshot() to do a poll loop after the image
snapshot is issued to wait until the image becomes available for use.

Change-Id: Ief8a3f6c32b56230f0509e658b14795149647ec3