Updated openstack/openstack Project: openstack/heat  1c95e4ff8664e4a6249c73db3a138e9da57de3c3
Make encryption script log exception and continue

The heat-manage update_params encrypt script is meant to be used
as a tool for encrypting parameters in database entries for older
stacks. For encrypting the database, the service may be required
to be taken offline; thus the script should be as robust as
possible. In particular, if the db entry contains malformed data,
instead of crashing, the script will log the error and continue
processing of further entries.

Change-Id: Id9c98ee87adf64f58ba25bb19269a38af71853d5
Co-Authored-By: Pratik Mallya <pratik.mallya@gmail.com>
Closes-Bug: #1488326