Updated openstack/openstack Project: openstack/glance  39c327fa6d345625d12cc56f7ad5f0a4ceba3d8f
Handle SSL termination proxies for version list

Return correct scheme in version URLs if service
behind an SSL termination proxy.

This is done by adding a new configuration option,
secure_proxy_ssl_header, which, when defined, makes
the wsgi application take the host_url scheme from
that header. By default, when this option is not
specified, there is no difference in behavior.

The intention is to configure any ssl-decrypting
proxy to set that header, so that glance-api knows
which protocol to use in the URLs in response.

This patch is largely based on the equivalent
nova patch: https://review.openstack.org/#/c/206479.

Partial-bug: 1558683

Change-Id: I9a9c0e42a6ad3c18d197f10095958b48d5cb879a