Updated openstack/openstack Project: openstack/cinder  6ae13613c452cb5a361af460554d6a7ce41898d6
Remove entry_points from setup.cfg

The keystonemiddleware entry_point and the oslo_concurrency,
oslo_db_concurrency and oslo_messaging entry_points are not
owned by Cinder so it is possible that Keystone and Oslo could
move or rename the list_opts methods causing our sample
configuration generator to break.

This patch removes the lines that register the entry_points since
they aren't actually adding anything to the sample list of
configuration options because the options are being added to the
sample file elsewhere.

Change-Id: I186d4d9e8e565dad836543b5f6a96af1a53d4a83