Updated openstack/openstack Project: openstack-infra/release-tools  bc44fdb612b8a3b4334a9cefce1664b0f269b1f3
improve check_library_constraints.sh to edit the constraints file

Update the script to find outdated constraints and update them. It
depends on having a copy of the right branch of the requirements repo
already checked out, but it clones the other projects itself and then
edits the constraints file.

Change-Id: Idfacbc937adc7c70fe4505b47fe380e850f1899f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>