Updated openstack/openstack Project: openstack-infra/project-config  b3b9825b7f1675a3e7196a6cc41297f5502f0118
remove sahara-bandit job

this change removes the sahara bandit job as it will be moved into the
pep8 test for sahara. this change is being made because the bandit tests
are now in a place to be made voting.

this change also upgrades the bandit integration job for sahara to
voting, addressing the TODO left there.

Depends-On: Ia0ad32666f85d289768f66199c088380ce363b54
Change-Id: I2f4a743b37c28264011bd9ffb230be0609102523