Updated openstack/openstack Project: openstack/nova  1ade1760f00fac8993d4085f9ebf8dd9b6467e0e
Populate instance_mappings during boot

We need to start populating the mappings of instance to cells. For now a
mapping with a null cell_mapping will indicate that the instance has not
yet been scheduled and that list/show requests should be honored using
the persisted RequestSpec/BuildRequest. That work will be a followup to
this one.

There's no need to leave instance_mappings loitering in the database if
there are no cells yet. So if a scheduled host does not belong to a cell
the instance_mapping will be removed. The instance will be mapped later
as part of the migration process to cellsv2.

Partially-implements: bp add-buildrequest-obj
Change-Id: I6f02e966d2cef85f98b4b64aedd8c352387af15d