Updated openstack/openstack Project: openstack-infra/project-config  c68aeeb988f2f670ee48d8becfb1f17bca683ca4
Update build-image.sh for new images, add EXTRA_ELEMENTS

This changes the base element for build-image.sh to ubuntu-minimal, and
adds the simple-init element, to make this match the images used in the
gate.

It also adds an EXTRA_ELEMENTS environment variable that users can use
to add elements to the image. For example, local-config is useful to put
the current user's ssh keys in the image.

Last, it documents the things I needed to install to get it to run
correctly on Ubuntu; this is likely not an exhaustive list.

Change-Id: I2506c6722b77bb2f861d08e7a6da453dc2bbbfc2