Updated openstack/openstack Project: openstack-infra/project-config  f6bd19645442de10e5ffe97ad3d5c02f23d1b87a
[new project] Create a OpenStack related project "daisycloud-core"

daisycloud is an installer which provides automated deployment and
management of OpenStack and other distributed systems.

daisycloud is an open source project (https://github.com/opendaisy)
and has been used widely by our customers already. To attract more
developers to join us, we would like to take the advantages of using
the OpenStack tool-chain and development infrastructure.

daisycloud-core is the core module project of daisycloud.

Change-Id: I10af17adb8510848ef6148527cc3cd7713104087
Signed-off-by: Zhijiang Hu <hu.zhijiang@zte.com.cn>