Updated openstack/openstack Project: openstack/neutron  978672305d44844cd7353aadbf5c531d156841c0
SG protocol validation to allow numbers or names

SG rule protocol provided is validated against the DB rules'
protocols for both number and name. The filter provided to DB
is modified so that it is queried for records with both the
protocol name and number, instead of exactly the type provided
with the input. The returned DB rule record's protocol field is
validated against the supplied SG protocol field for both name
or number.
This way, user is still allowed to enter protocol name or number
to create a rule, and API compatibility is maintained.

Change-Id: If4ad684e961433b8d9d3ec8fe2810585d3f6a093
Closes-Bug: #1215181