Updated openstack/openstack Project: openstack/neutron  54653c4e593647c80ccef4419bd41ac71ad0fb63
Use context manager in bind NetworkDhcpAgent

In bind() method of dhcp_agent_scheduler for transaction was not
used context manager. This is not correct as if another exception
appear, not DBDuplicateEntry, that is caught, current transaction
will hang.

Current change is a simple refactoring, the code works as it does
prevously. Existing unit tests
* test_schedule_bind_network_multi_agent_fail_one,
* test_auto_schedule_network(Network already scheduled)
cover possible issues.

Closes-bug: #1557546

Change-Id: Ieb77738e065b997e0ab65afdc1f3bdbfb8f13fef