Updated openstack/openstack Project: openstack/governance  d0b6a611223eb6b20bd2b36f2dbbd0e83eff40e8
Add searchlight-ui to searchlight

The Searchlight content from Horizon is moving from a patch to a
plugin.

At the horizon mitaka mid-cycle, we decided to move this to a plugin
for the Mitaka release. It was not possible to do so until
several patch dependencies merged on horizon master. Those
have now merged, so we are able to put this into its own repo.

Change-Id: I035c479102200cce38511482f433b030b066584d
Depends-On: I528537a525a54db3887688ab49b33983403dd2ca