Updated openstack/openstack Project: openstack/heat  ad72957108a0bb07565ef68c91d5d0745287f8d6
Increase timeout of WaitCondition due to ConcurrentTransactions

Currently we have retry logic with random period of delay.
There is a chance, when we have several re-try attemts with maximum
timeout, so it can be a root cause of floating fails, when we have not
enough time for getting all signals.

Also decrease number of signals from 30 to 25, due to small risk, when
one of signals goes to ConcurrentTransaction error.

Note, that all these changes are made in corresponding functional test
and don't affect main code.

Change-Id: I9ca85c4378d0f62e316f4f40ff73597ffaaabadf
Related-Bug: #1497274