Updated openstack/openstack Project: openstack/api-wg  c3a5ba0b086cc83c46204ebf346a085a53a54b8e
Clarify the guideline proposal process

The steps in the proposal process were interleaved in a way
that made following them confusing. This change extracts the
definitions of terms from the actual process and simplifies the
process.

For the most part the spirit of the staged review process is
maintained but the steps required to reach out to the community
are clarified: Where there is consensus we involve just the CPLs.
Where there is not we reach out to the rest of community. I'm not
certain is the correct way to go. It's a compromise between keeping
things moving with minimal friction and including the broader
community as much as possible.

My hope is that this will help to clarify the term "frozen" and
when freezing is to be done.

This will probably require some iteration.

Change-Id: I416ff1dd35249674fb27f9dd7625786e52001556