Updated openstack/openstack Project: openstack/api-site  cc5ebf3b9e959d55b4cd1d040fa956feabe4b9de
Put in example block_device_mapping structure

The documentation on the block_device_mapping_v2 structure doesn't
really explain much. This provides a sample structure so that it at
least gives a starting point for people. It would be best if this was
expanded on further, but it's a start.

Closes-Bug: #1553152

Change-Id: If559035fbcf3d4c79bae5672abf6718915917d72