Updated openstack/openstack Project: openstack/horizon  a0173c7469a075e02577f35a199d18a0676731df
Port list check is missing for LBaaS Add Member

The "Add Member" workflow calls neutron to retrieve
a port list. The exception is handled, however there
is a case when port list is empty and therefore
a member cannot be created. This patch set adds a check
and a corresponding failure message.

Change-Id: I86493acb47e15ab9348fa5029852e873b64358aa
Closes-bug: #1521394