Updated openstack/openstack Project: openstack/keystone  b66564e0938470544cd177f30966c14443022dd5
The name can be just white character except project and user

The reason for this bug is that the name is not validated when create
 role, domain and groups. In the V3 version we use jsonschema to
validate them. So I changed the code of schema files to validate
the name.

When I wrote the test code, I found that there was a lack of some tests
 in the file and then I added to write the test code.

Change-Id: Ie7bd79bede3b405b3798bdb584d2fcab66104317
Closes-Bug: #1537697