Updated openstack/openstack Project: openstack/tripleo-heat-templates  717d949ddfe128c84239418c5559464172ec236f
Don't restart Pacemaker-managed services from Puppet

We need to reload/restart services on updates/upgrades to apply any
config changes, but restarting services managed from Pacemaker from
Puppet causes problems.

For now we no-op the restart and rely on the catch-all restart after
Puppet phase.

In the future we should have a service provider for pacemaker resources
that will be using pcs. We still might have to restart services outside
Puppet due to cluster-wide orchestration issues, but we might be able to
do the restarts selectively rather than restart everything.

We also no-op the start/stop commands to be safe, as it also doesn't
make sense for Puppet to try start and stop those services when it
doesn't have knowledge about Pacemaker.

Change-Id: I95e21e10471cd7575f28c095c48150325f1414b3
Closes-Bug: #1562922